replicaset: introduce map_call()
There is a bug about master instance's bucket GC fiber not
respecting RO refs on replicas. The idea of a fix - make the
bucket GC procedure consult each replica before marking a bucket
as GARBAGE. Why this way - see motivation in the main commit.

Consulting each replica means map-reduce. The present commit
introduces replicaset:map_call(). It is like router.map_callrw(),
but works on all replicas of a single replicaset instead of
masters of all replicasets.

It is going to be used in the main commit.

Needed for #173