test: fix flaky router/router2.test.lua
The part of the router/router2 tests changing discovery mode to
different states. During reconfiguration, which is invoked after
every of these changes, previos discovery fiber is killed and new
one is created.

The problem is that we check the statuses of the fibers without
any waiting and sometimes there was no time for fiber to be killed.
This happens due to the fact that fiber's cancellation is performed
asynchronously: it's not instant.

Let's wait until the status of the last tested fiber is not dead.
By this point all other fibers will already be dead anyway.
Moreover, the last fiber, which is executed only once, can be
already dead by this time. Let's check if it's done or still working.

Closes #358