gc: do not check local buckets via netbox
Bucket GC makes a map request on all nodes in the replicaset to
check which buckets are eligible for turning into GARBAGE.

Going to the own local instance via netbox of course makes no
sense. The buckets are checked locally both before and after the
map request anyway.

The patch optimizes GC so that it won't visit self via netbox
during the map call.

Part of #173