gc: wait replication before checking SENT buckets
GC goes to replicas to check if a SENT bucket can be deleted. But
it can happen that this call via netbox was faster than the
replication. Some replicas still can have the bucket SENDING.

That triggered 5 seconds GC backoff. That was not a problem for
GC, but it was for map calls. Router's map_callrw() had to wait
until GC woke up after the backoff and finally deleted the bucket.

All that time the pending map_callrw() was not only just waiting,
but also disrupting rebalancing. Because the "bucket-move vs
storage-ref" scheduler tries to give time both to moves and refs
fairly.

The patch makes bucket GC synchronize with replicas before
checking their buckets. It is done per-batch. Doing it just once
in the beginning of _bucket space iteration wouldn't be enough,
because new SENT buckets not covered by that sync would keep
appearing (if the rebalancing is still ongoing).

Part of #173