test: decrease wait_timeout
Currently `wait_timeout` is set to 120 seconds, which is also a default
value for luatest. The problem is the fact, that any timeout loop, if it
hangs, will be executed until the process won't be forcefully killed by
luatest. In such case result file is not created and the programmer
can't even see, what happened.

Let's set `wait_timeout` to 50 seconds, which seems to be pretty enough
for any test to be completed.