test: fix flaky router/router
The part of the test router/router inserts a tuple to the space
of the master, invokes 'vshard.storage.sync' on the same master
and executes 'vshard.router.callro', which routes the request to
the replica of the corresponding replicaset.

The problem was that sometimes this replica didn't have needed
tuple and returned null, which wasn't the value the test expected.
It was caused by the fact, that sometimes replication didn't have
time to happen and 'vshard.storage.sync' didn't check vclocks of
all replicas in a replicaset.

The purpose of the 'vshard.storage.sync()' is to wait until the
dataset is successfully synchronized on all replicas. The work of
this function relies on the downstreams from 'box.info.replication'.
These fields are not 'nil' only if the instance knows about the
other ones, which follow her. In the previous implementation of the
'wait_lsn()', which is used in 'sync()', the function returned true
even if there was no 'downstreams' at all, which means that it used
to said that the data is fully synchronized, without checking
vclocks of all replicas.

The fix is to make the 'wait_lsn()' function return true only if
there's downstreams to all replicas and vclock of the current replica
is less or equal than any of replicas vclocks.

Closes #366