ci: add linters
This patch introduces running luacheck and checkpatch on CI.

It uses the lastest version of luacheck available on luarocks.
We should add ignoring of warning in 'version_test' file as this
version induces us to replace all comparison signs to be
positive (without using 'not' boolean expression). Moreover, due
to the recent changes mutating of the ivshard global variable was
added, which should also be ignored.

Checkpatch should ignore NO_CHANGELOG and run only on PR in order
not to make CI red when pushing directly to the master branch.

Closes #369

NO_DOC=ci
NO_TEST=ci