storage: do paranoid checks of _bucket and refs
Previously updates of _bucket were not validated. One could drop a
bucket which had refs, or set an invalid bucket status, or do
something equally dangerous.

This commit makes _bucket:on_replace() trigger perform checks.
Like SQL CHECK constraints in core, but on_replace() also works at
1.10.

From now on the illegal changes are banned. Can't set GARBAGE
state for a bucket having ro refs; can't start RECEIVING a bucket
if locally there is a somehow not deleted old bucket_ref for this
BID; can't make bucket SENDING if it has rw refs; etc.

The patch also forbids to truncate _bucket. Because checking all
the bucket_refs that all their ro and rw counters are zeros might
take unpredictable time and yielding on _truncate space update is
not safe without MVCC (and frankly with it as well).

The protection can be turned off if something got broken and needs
manual intervention. It is not documented though, because is not
supposed to be used in application's code.

There was an option to make on_replace() also check the bucket
state transitions, ensure that a bucket status update is one of
these:

    * none -> RECEIVING
    * RECEIVING -> GARBAGE
    * RECEIVING -> ACTIVE
    * PINNED <-> ACTIVE <-> SENDING -> SENT -> GARBAGE -> none

and it worked in a draft patch, but it requires changes in almost
every test. For example, there are a lot of tests which simply
drop buckets one by one or re-create them. They would have to
create them as RECEIVING and turn into ACTIVE. Or turn the
protection off which would reduce the tests' usefulness.

Part of #173

NO_DOC=internal changes