test: fix flaky storage/storage.test.lua
Sometimes it was failing (both with and without output) with an
error in .reject file that `vshard` variable wasn't defined after
storage_1_a restart.

The reason was that the restart was done with `wait=False` param.
But there is no reason for it. Maybe there was before, but now
the waiting is certainly fine.

Without the waiting it could happen that box.cfg is called and
listen is done, but box.cfg didn't end yet and
`vshard = require('vshard')` apparently isn't done.

Closes #378

NO_DOC=bugfix