test: fix flaky storage-luatest/bucket_triggers
Currently garbage collector works during checking the bucket
transitions. Sometimes it turns SENT buckets, we manually created,
into GARBAGE and the test fails to check the right error message.
The solution is to disable gc during this test.

Some of the tests're also disabled in this commit due to the bug
in mvcc tarantool/tarantool#7977.

Closes #386

NO_DOC=bugfix