storage: make on_master triggers to return value
Currently `on_master_enable/disable()` doesn't return anything.
However, it might be useful for a user, which closures as triggers,
in order to receive the function, which was set. Moreover, users
might also want to get the list of installed triggers, which isn't
possible now.

Unfortunately, for some reason triggers, created from the Lua API,
don't return function. It'll possibly be fixed in the nearest future.
So we should update tests and suppress the output value of these
triggers as the function pointer difffers every time the test run.

Part of #372

NO_DOC=<undocumented part>