util: fix creating nameless fibers
The problem is the fact that a reloadable fiber starts before the name
is set. So until the first fiber yield its name is 'lua', which is the
default value for `fiber.create` or `fiber.new`.

If the user greps logs in order to get info about some specific fiber
he will probably miss the beginning of these logs as they are
indistinguishable.

Let's use `fiber.new` and not `fiber.create` in order to initialize the
fiber. It won't start before context switching is performed. This way
we will have time to set the fiber's name.

We should also wakeup this fiber and yield the current one in order
to prepare and start the fiber which was created with `fiber.new`. This
is done so as not to change the current behavior of the
reloadable_fiber_create function.

NO_DOC=bugfix