service_info: introduce the module
This patch introduces the new module, the purpose of which is
to save statuses, activities and errors of any background service.

It's mostly done in order to ease the luatest testing of the background
fibers. So, this commit also adds helper functions for working with
service_info module.

It can also used for monitoring of the vshard's services and shows
whether any errors occurred and what this service is doing right now.

Part of #107

NO_DOC=internal