test: fix upgrade.test.lua not working on 3.0
The test tried to call vshard.storage.cfg{} second time on 3.0
using an old vshard version. It won't work because the old
non-first cfg tried to compare box.info.cluster.uuid with the
replicaset UUID in the config. But box.info.cluster.uuid is nil in
3.0. It was moved to box.info.replicaset.uuid.

The patch makes the upgrade test use the first commit compatible
with 3.0 as the basic version, if the Tarantool executable is 3.0.

Note that on versions < 3.0 the vshard upgrade still works fine,
hence on old executables the basic version is kept as is.

Follow up #402

NO_DOC=test