test: fix flaky rebalancer/rebalancer.test.lua
Currently the test hangs sometimes on the line, which waits for the
message "Rebalance routes are sent" to appear in logs.

According to the logs this message appears, when hang happens, but it
is located before the bunch of "aa..a", which means, that rebalancer
managed to send routes before waiting actually started.

This happens due to the fact, that rebalancer is waked up before
waiting. Let's not do that.

Closes #360

NO_DOC=testfix