test: use dofile() for configs instead of require
The main reason of this change is that we're going to introduce a new
built-in module `config`. See [1] for details.

Introducing a new module in the global namespace is legal according to
our compatibility rules. Quoted from [2]:

> Adding a new built-in module or a new global value is considered as
> the compatible change.

Moreover, in the real-world applications configs may be changed and
reloaded, so they likely will be loaded using dofile() or loadfile()
without require()'s caching.

[1]: https://github.com/tarantool/tarantool/issues/8724
[2]: https://github.com/orgs/tarantool/discussions/6182

NO_DOC=adjust tests for the new tarantool version