storage: allow to choose rebalancer location
Currently people struggle to find the rebalancer location because
it is unclear where it is spawned automatically (on the master
with minimal uuid).

The patch allows to choose in the configuration where the
rebalancer should work and whether it should work at all.

Closes #278

@TarantoolBot document
Title: vshard: cfg to choose rebalancer mode and location

There are 2 new config options:

- Root option 'rebalancer_mode'. It can have one of the following
  values:
  - 'auto' - works like now. The rebalancer is selected as the
  master with the minimal UUID;
  - 'off' - rebalancer won't be active at all;
  - 'manual' - need to manually mark the node where the rebalancer
  should be spawned. The selection happens in the config.

  Default is 'auto'.

- Instance option 'rebalancer'. It can be true or false. At most
  one instance in the config can have it set to true. If the
  rebalancer mode is 'manual' and none of the instances have this
  flag true, then the rebalancer won't be started anywhere. This
  option is ignored when the mode is not 'manual'. Default is
  false.