storage: introduce master_call() wrapper
It wraps replicaset:callrw(). The purpose is to handle the
NON_MASTER error in a single place. It is going to become much
more likely to get after the next commits which will enable master
discovery at the storage level.

It will be important to notice that the node, which was thought to
be the master, is not in that role anymore, and a new discovery is
needed.

Part of #430

NO_DOC=internal change
NO_TEST=covered in next commits