test: extract service_info into a separate file
It is taking too long time (> 30 seconds). The commit moves it to
a separate file so as the more generic test storage_1_1 wouldn't
be too slow. The latter is going to get new test cases and it is
easier to debug them when the info-test is in another place.

Needed for #429

NO_DOC=refactoring