rebalancer: enable it on auto-master replicasets
Previous commit added support for full master discovery on the
storages, but the rebalancer didn't work on replicasets having
auto master enabled for themselves. Now it works on those too.

The priority is firstly given to manual masters. If none are found
in the entire cluster, then the rebalancer appears on the
auto-assigned master with the minimal replicaset UUID among all
the other auto-masters.

Closes #430

NO_DOC=documented later with new rebalancer options