storage: fix too early drop of SENT buckets
SENT bucket could be deleted by GC before bucket_send() is
completed. It could lead to a problem that bucket_send() would
notice that in the end and treat as an error. The bucket should
not be deleted while its transfer is still in progress.

The bug is fixed now because it started manifesting frequently in
new tests for the next commits.

Closes #433

NO_DOC=bugfix