replicaset: rename is_auto_master to is_master_auto
The old name looked like it meant the object itself is an auto
assigned master. But in fact it is a replicaset flag which means
the master role should be auto-discovered.

The patch makes the name clearer. The reason for it now is that
the flag is going to be used more in the next commits.

Part of #429

NO_DOC=refactoring