rebalancer: return nil in rebalancer_apply_routes
The function used to return 'false' on error. But all functions
which can fail are supposed to return nil on error (exception on
critical errors such as misusage).

The patch makes the function conform with the rules. This is done
as a preparation before the function would start returning another
error in a next commit.

Part of #429

NO_DOC=not visible
NO_TEST=already covered