storage: use util.fiber_cond_wait everywhere
A few places in storage code still used fiber_cond:wait()
directly. It is not safe for negative timeouts and is too easy to
forget to convert an error to vshard format. The patch replaces
those usages with util.fiber_cond_wait.

This was noticed while fixed another bug related to sleeps and
fiber cancellation.

NO_DOC=not visible