storage: move master and rebalancer role commit
It was done after schema upgrade, but now is done before. It makes
sense because master commit should enable read-write mode which is
needed for the schema changes.

It worked so far, because in manual master config the read-only
cfg update is done at the box.cfg stage. I.e. before the schema
update.

But soon the master role will be allowed to turn on and off
automatically. And then the old order wouldn't work anymore.

At the same time not everything can be done before the schema is
installed. Because some services, like recovery and bucket GC do
need the schema (`_bucket` space at least). The solution is to
make those services depend on whether the first configuration is
performed. They are started only after that.

Part of #429

NO_DOC=refactoring
NO_TEST=refactoring