storage: introduce is_master flag
There is this_is_master() function, but it wasn't used in all
places, and it used to rely on the config. Soon the config won't
be telling who the master is when the role is set to be automatic
in next commits.

The patch introduces a flag is_master which reflects the current
actual state, not the config. For now it is only managed by the
config, but in the next commit it won't be so anymore.

Vtest also now won't rely on the config. It uses the actual master
state.

Part of #429

NO_DOC=internal change