cfg: introduce extract_vshard and extract_box
Those functions replace the old split() function. It is needed
mainly for tests which are going to pass read_only options
per-instance in the global config. But it also looks like an
actually useful feature even outside of tests.

In tests that is needed for auto-master cases. There is already
one place - auto_master_2_test.lua, but for that one the patch in
vtest was enough and helped only for bootstrap. The current patch
allows to reconfigure the instances after bootstrap via their
own private config sections.

Needed for #430

@TarantoolBot document
Title: vshard: per-instance box options in the config

Previously the config could include box options only on the root
level like so:

```Lua
{
    -- VShard options:
    sharding = ...,
    weights = ...,
    bucket_count = ...,
    ...
    -- Box options:
    replication_timeout = ...,
    memtx_use_mvcc_engine = ...,
    net_msg_max = ...,
}
```
Those box-options are propagated to `box.cfg{}` on each instance
using this config. But there was no way to pass different options
for difference instances via one config.

Now it is possible via their config sections in `sharding`:
```Lua
{
    sharding = {
        [replicaset_uuid] = {
            replicas = {
                [replica_uuid] = {
                    replication_timeout = ...,
                    net_msg_max = ...,
                    read_only = ...,
                }
            },
            ...
        },
        ...
    },
    ...
    -- Box options:
    replication_timeout = ...,
    memtx_use_mvcc_engine = ...,
    net_msg_max = ...,
}
```

For each instance its resulting `box.cfg` is merged from the root
box options and from the instance's section in `sharding`. If some
options are specified in both places, then the priority is for the
replica-local options.

For example:
```
{
    sharding = {
        [replicaset_uuid] = {
            replicas = {
                [replica_uuid] = {
                    replication_timeout = 100,
                    net_msg_max = 200,
                }
            },
        },
    },
    replication_timeout = 50,
    txn_timeout = 300,
}
```
The resulting `box.cfg` config for the given replica will be:
```Lua
{
    replication_timeout = 100,
    net_msg_max = 200,
    txn_timeout = 300,
}
```