storage: introduce full master discovery
Until now only master discovery of own replicaset was supported on
the storage level. But now the storages can find masters of other
replicasets.

The design is similar to the router's master discovery, except
that it is not proactive. Storages rarely need to talk to each
other which makes pointless to constantly maintain all the
connections and re-check the instances' roles to keep the master's
location info up to date.

Instead, the discovery is reactive. It only works for the
replicasets which are being used for any calls right now.

Part of #430

@TarantoolBot document
Title: vshard: master discovery on storage level
Routers do already support `master = 'auto'` setting at the
replicaset level in the config. It makes them able to
automatically find the master instance location in the so
configured replicaset.

On the storages it wasn't supported at all. But now it is. The
storages can specify in their configs the same option
`master = 'auto'` for their own replicaset and for any other
replicasets too.

If the storage has `master = 'auto'` for its own replicaset, then
it will constantly monitor `box.info.ro`. When it becomes `true`,
the instance considers itself a replica. When false - considers
itself a master. It also means that such automatic masters will
start and stop all needed background fibers automatically as well,
depending on the actual state of the instance, not on the config.

However it is worth keeping in mind that this option doesn't mean
master election. It only means master **detection**. I.e. if this
option is used, then the user's code is supposed to assign the
masters using `box.cfg.read_only` or any other way.

If the storage sees `master = 'auto'` for any other replicasets
than its own, then it will be able to automatically discover their
master instance for internal operations, such as bucket recovery
and rebalancing.