cfg: introduce type_descriptor
Type validation in cfg module used to be done by just value native
type. It is clearly not enough for enum type which is currently
used only twice (for cfg.sharding.replicaset.master and
cfg.discovery_mode) and is hacked via the check() validator, but
soon more enums will appear (for user management and rebalancer
mode).

The patch prepares the code for the appearance of the enum type.
For that the type validation needs to have a context, not just
type name.

Needed for #435

NO_DOC=refactoring
NO_TEST=refactoring