test: introduce test_user_grant_range cfg option
This is a vtest option which is used only by vtest and is not
propagated into vshard.

The purpose is not to grant 'super' privileges to the vshard user
when it is not needed. It allows to ensure that vshard can work
fine with fully automatic user management without any external
intervention (such as via vtest).

It will be used in a next commit for testing of non-automatic
user management.

Needed for #435

NO_DOC=internal