storage: propagate errors of services
Currently it's difficult to debug services, as they don't throw an
error if functional part failes. This leads to the log messages
about a module being reloaded without explanation of the reason.

This is caused by the fact, that functional part of every service is
pcalled, return status is not checked. Let's throw a returned error,
if pcall of a service fails.

Part of #446

NO_DOC=bugfix
NO_TEST=<following commit>