storage: fix assertion failure in conn_manager
conn_manager_collect_idle_conns is used in order to close connections,
which have not been used by a storage for a period of time. Currently,
it has an assertion, which checks, that node never connects to self.

However, storage establishes connection to self, e.g. during
downloading of bucket states for rebalancer. As this is expected
behavior, this commit deletes this assertion.

Closes #446

NO_DOC=bugfix