storage: extract schema control into a new module
vshard.storage.schema is a new private module for dealing with the
schema: upgrade, downgrade, and all the helpers.

This is needed because the schema business is going to get more
complex soon, in scope of configurable schema management mode.
Also a potential side-feature will be that the external code will
be able to bootstrap/upgrade vshard schema without doing
vshard.storage.cfg(). This is needed for `config` module in core
3.0.0.

Part of #263
Needed for #435

NO_DOC=internal