router: improve master connection parallelism in map-reduce
This is useful for RW map-reduce requests which need to send
multiple network requests in parallel to multiple masters.

In-parallel means using is_async netbox feature. But it only
works if the connection is already established.

Which means that the connection establishment ideally must also
be parallel.

NO_DOC=internal
NO_TEST=already covered