storage: introduce exports and their log
The new module vshard.storage.export_log contains a list of all
exports for each vshard schema version. It is used by vshard to
deploy the exports, and is going to be used by the `config` module
in the latest core 3.0.0.

The new module vshard.storage.exports provides some helpers for
using the export log. For instance, to deploy exports of one
specific version.

Part of #435

NO_DOC=internal, those modules are only for config core module.