storage: introduce schema_management_mode
The new option allows to turn off automatic creation and upgrade
of functions, users, and privileges of all sorts. For that the
mode has to be specified to 'manual_access'. The default value
'auto' makes vshard behave like before this option's introduction.

Note that the option is intentionally not documented so as not to
encourage people to use it. The best approach is to let the schema
to be maintained fully automatically.

Closes #435

NO_DOC=internal