router: relax checks in auto disable
Router's API is protected until the configuration is done, as accessing
router's functions at that time is not safe and can cause low level
errors.

Let's relax checks in this protection. The thrown error messages are
incorrect from now on: router can be started on unconfigured
instance, when box_cfg_mode is 'manual'. box.info.status in such case
will be 'unconfigured', not 'running', we shouldn't rely on it either.

NO_DOC=<documentation is correct>