test: introduce infrastructure for named config
This commit allows vtest to create named config, wich will be introduced
in the following commits. If identification_mode is 'name_as_key', then
names are used instead of UUIDs as keys, replica/replicaset.name is not
specified, as it's forbidden, UUIDs are set as replica/replicaset.uuid.

The commit also introduces new server API: instance/replicaset_name(),
which is meant to be used in tests for Tarantool >= 3.0.0.

Part of #426

NO_DOC=internal