error: more generic argument names
In the following commits it'll be allowed to use names instead of
UUIDs. This patch refactors error module, changing the name of
arguments, accepted by errors: from now on it may be either UUID or
persistent name, so we cannot say, that this is always UUID, as it's
done now.

The alternative would be to rename arguments according to the
configuration used on the fly. It's not selected, as it complicates
expluatation.

Note, that this is breaking change, if user's code relies on UUID
arguments of returned errors!

Part of #426

NO_DOC=refactoring