test: fix flaky failover/failover.test.lua
Router pings replicas and if they cannot respond in
failover_ping_timeout, it detaches connection from this replica.
The test checks, that connection is not closed, when long response is
done, even if failover_ping_timeout is small.

However, when the connection is detached, it can be garbage collected,
which leads to 'Connection closed' error. Let's stop Lua GC for the
connection by creating reference to it.

Related to tarantool/tarantool#9629

NO_DOC=test