test: update `async_no_yield` test from `vconnect` test group
After tarantool/tarantool#9489 is fixed, we can update the test to reflect
the correct behaviour. However, we still need to test the incorrect
behaviour for older versions of Tarantool to which the bugfix was not
backported to.

Due to tarantool/tarantool#9489, we have deadlock: either Tarantool's
integration tests fail with vshard or testing against the Tarantool master
version fails in vshard. Let's temporarily disable the assertion that
tests against the correct behaviour in newer Tarantool versions and causes
failure.

Part of #456

NO_DOC=<test>