storage: expose uri to info
This commit exposes URI of the current replica to vshard.storage.info().
Crud parses URI in order to extract user of the current replica, which
is used for automatic stored procedure grants setup. However, after
introducing of auto-master on storages URI of the master is not shown in
vshard.storage.info() and crud has to use internal fields to access user
name.

Closes tarantool/vshard#467

NO_DOC=<minor change>