router: fix master discovery stall
When router master discovery is enabled, it is supposed to find
the current master instance automatically for the corresponding
replicaset. It worked fine as long as the master switch is
graceful and the old master is still visible to the router when it
is demoted.

But if the old master was disconnected without any notice, then
the router wouldn't notice the new master. It wouldn't even try to
find it really. Instead, it would keep bumping into the old not
available master trying to check if it is still in its role.

The problem was that the master search function didn't try to
check other nodes as long as its knows last master location for
the given replicaset.

Another related issue was that the router's master search
procedure was only running with a hardcoded period of 5 seconds.
If the current master gets disconnected, in average it would take
~2.5 seconds to notice that even if there is a load right now.

The patch addresses these issues by trying to check all nodes if
the last known master is disconnected, and by speeding the master
search up if a request comes and notices that the master is
disconnected. Luckily, it is easy to do using `on_master_required`
replicaset trigger introduced for storages some time ago.

Part of tarantool/vshard-ee#4

NO_DOC=bugfix