replicaset: introduce stateless balancing for read-only requests
This commit introduces balancing for requests, where balance mode is not
set. The motivation for this change is that we should make our best to
succeed with call, which is important for mission critical systems.

When balance mode is set, replicas are balanced between requests,
consequent calls won't go to the same replica. Balancing is done
according to the round-robin strategy, weights doesn't affect such
balancing.

However, when balance is false, callro and callre are still balanced
now. But this balancing happens only in the scope of the call,
consequent calls go to the same, most prioritized replica firstly.
Retry happens, when network error happens (e.g. timeout).

During such balancing callro method doesn't distrinuish master and
replica. If master has the highest priority according to config, then
callro will go to master firstly, and only after that to replica. callre
method, on the other hand, will firstly try to access all replicas, if
all requests fail, then we fallback to master.

Closes tarantool/vshard#484

NO_DOC=<ticket was already created>