test: move test_named_config_identification to a separate file
The problem with this test is that it sets names, which is not easily
revertable. Consequently, if it's located in a big test suite, like
storage_1_1, we cannot rely in other tests either on the presence of
the names or on their absence, which is very unconvenient.

Let's move the test in the separate file, so that in storage_1_1 we
always know, that the names are not set.

NO_DOC=test