storage: introduce a way of skipping bucket status state
After rebalancer failure we need to manually modify the state
of buckets, delete or create them. This is very difficult to do
without this error injection, since it requires manually disabling
triggers on _bucket and updating cache after change.

Let's introduce the error injection so that we have a handy way to
recover cluster after failures.

Closes tarantool/vshard#489

NO_DOC=internal