storage: introduce ref.check()
It ensures the ref is still in place. A read-only operation. It is
going to be used in the future commits about partial map-reduce.
Router will be going potentially more than once to some storages
and at all times it would ensure the ref is still in place.

NO_DOC=internal