replicaset: do not update replica_up_ts on connect
Before this patch `replica_up_ts` was updated in on_connect netbox
trigger. This variable is used by the failover router service to
figure out, when it's time to check, whether priority of the replicas,
which are higher in prioiry list, can be increased.

It was updated in on_connect trigger so that failover doesn't try
to increase priority, when it's not needed. The problem with this
update is that it requires `replicaset` module to be in `conn` and
this requires, that every connect is done above replicaset object.

This doesn't seem right, as replica object should have enough
info to connect without need for replicaset object. In the
following commits failover will be moved to the replicaset module
and it will be divided to several services: one for replicaset,
other for replica. The latter will do pings, but since ping
requires replicaset module to detach broken connections, it cannot
be done on top of replica object only.

This commit removes this small and unnecessary optimisation, but
allowes to connect replica, without using replicaset object.

This commit doesn't include doc request, because only call methods
of replicaset are documented, all other methods can be considered
as internal API and don't require documentation.

Needed for tarantool/vshard#505

NO_DOC=<not needed>