router: drop ERRINJ_FAILOVER_CHANGE_CFG
In the following commits failover will be moved to replicaset.
Failover step will be run on top of every replicaset and
situation, when one of them becomes nil is not possible anymore.

For now only error injection is dropped, code, which process
situation, where rs == nil will be dropped, when failover
is moved, as we cannot leave vshard broken after commit.

Needed for tarantool/vshard#505

NO_DOC=internal