storage: move master search to replicaset
This commit moves the remaining part of connection manager
from storage to the replicaset module. This part searches
for master, when someone is requesting for it, this is
called 'lazy' mode of the master search.

The commit also introduces 'aggressive' mode, which
will be used by the router in the subsequent commit.
In aggressive mode master search service searches for
master every step, even if nobody requested.

Needed for tarantool/vshard#505
Part of tarantool/vshard#263

NO_DOC=internal