replicaset: replication state check is disabled, when master is down
Consider the following situation: replicaset consists of 2 nodes,
master and replica, master is down. Since we don't check, that
there's node with alive replication in replicaset, we'll lower
the priority of the replica, master will become the prioritized one.
Replica still works. This way we'll always make request to master
first and only then to replica, which is not the expected behavior.

The situaton becomes much worse with implementation of
tarantool/vshard#505, since from now on all replicas will be skipped
during balancing, if master is down and fallback to master will also
fail, even read only request is not possible at all.

Let's check, whether we should lower priority of the replica
by checking health_status of all nodes in the replicaset.
It's worth mentioning, that replicaset_failover_step may
notice unrelevant state, when e.g. all replication checks of
the replicas already failed, but ping to master hangs due to
huge failover_timeout and it's still in YELLOW state.

Let's also wait for synchronization of the replicasets in the test,
since otherwise we may try to access the space, when schema has not
been replicated yet in other tests.

Closes tarantool/vshard#511
Needed for tarantool/vshard#505

NO_DOC=bugfix